+++ octopus/src/main/kotlin/com/koupper/octopus/annotations/AnnotationsResolvers.kt
@@ -1,6 +1,6 @@
 package com.koupper.octopus.annotations
 
-import com.koupper.octopus.logging.LogSpec
+import com.koupper.logging.LogSpec
 
 object LoggerAnnotationResolver : AnnotationResolver {
     override fun prepare(
@@ -11,7 +11,7 @@ object LoggerAnnotationResolver : AnnotationResolver {
         context: String
     ) {
         val level = (annotationParams["level"] ?: "INFO").toString()
-        val dest  = (annotationParams["destination"] ?: "file").toString()
+        val dest  = (annotationParams["destination"] ?: "console").toString()
 
         baseParams["logSpec"] = LogSpec(
             level = level,
@@ -35,7 +35,7 @@ object JobsListenerAnnotationResolver : AnnotationResolver {
 }
 
 val annotationResolvers: Map<String, AnnotationResolver> = mapOf(
-    "Export"       to object : AnnotationResolver { // Export no prepara nada
+    "Export"       to object : AnnotationResolver {
         override fun prepare(sp: String?, bp: MutableMap<String, Any>, ap: Map<String, Any>, s: String, c: String) {}
     },
     "Logger"       to LoggerAnnotationResolver,